==> analyses/molecular-subtyping-LGAT/00-v17-LGAT-select-pathology-dx.Rmd <==
---
title: "Select pathology diagnoses for inclusion"
output: 
  html_notebook:
    toc: TRUE
    toc_float: TRUE
author: Jaclyn Taroni for ALSF CCDL
date: 2020
---

## Background

Previously, samples were included for subtyping in this module if they were labeled `LGAT` in the `short_histology` column.

In an upcoming release, `integrated_diagnosis`, which can be updated as the result of subtyping, will be used to populate the `short_histology` column (see [#748](https://github.com/AlexsLemonade/OpenPBTA-analysis/issues/748)).

Thus, molecular subtyping modules need to be upstream of `short_histology` and use the `pathology_diagnosis` and `pathology_free_text_diagnosis` fields.
This change for this module is tracked in [#753](https://github.com/AlexsLemonade/OpenPBTA-analysis/issues/753).

We have explicit instructions for inclusion [here](https://github.com/AlexsLemonade/OpenPBTA-analysis/issues/753#issuecomment-697008356); we include this notebook in the interest of documenting and exploring the changes to the module in the code base itself.

## Set up 

```{r}
library(tidyverse)
```

### Directories and files

We're going to tie this analysis to a specific release (`release-v17-20200908`).

```{r}
data_dir <- file.path("..", "..", "data", "release-v17-20200908")
histologies_file <- file.path(data_dir, "pbta-histologies.tsv")
```

## Read in data

```{r}
histologies_df <- read_tsv(histologies_file)
```

## Explore the pathology diagnoses

### `short_histology == "LGAT"`

In the current histologies file, if we filter based on `short_histology` as we did originally, what is in the pathology diagnosis fields?
Note that some of the `short_histology` values will have been altered based on earlier subtyping efforts.
(That's why we're doing this!)

```{r}
histologies_df %>% 
  filter(short_histology == "LGAT") %>%
  count(pathology_diagnosis) %>%
  arrange(desc(n))
```

`Low-grade glioma/astrocytoma (WHO grade I/II)` certainly is an expected result!
During an earlier subtyping effort, one sample that was originally `HGAT` was reclassified due to the presence of a BRAF V600E mutation.
`Subependymal Giant Cell Astrocytoma (SEGA)` is also expected, but is not accurate - this will be changed in an upcoming release.

We're also going to include `Ganglioglioma` specimens for subtyping going forward, but that was not done previously and we should not expect to see it above.

```{r}
histologies_df %>%
  filter(str_detect(pathology_diagnosis, "Ganglioglioma")) %>%
  count(pathology_diagnosis) %>%
  arrange(desc(n))
```

Of these, we only want `Ganglioglioma` and `Ganglioglioma;Low-grade glioma/astrocytoma (WHO grade I/II)`. 

## Pathology diagnosis strings for inclusion

We need to include LGG samples and ganglioglioma samples, but these are often combined with other diagnoses in `pathology_diagnosis` and separated with a `;`.
If we were to use exact matching and the content/interpretation was unchanged, but the order of the diagnoses changed for whatever reason, this would break.
So we're going to go the string detection route.

```{r}
include_path_dx <- str_to_lower(
  c(
    "Low-grade glioma/astrocytoma",
    "Ganglioglioma"
  ))
```

There were no LGG in the tumors diagnosed as _only_ `Other` as of right now, so we don't need to look at `pathology_free_text_diagnosis`.

We do need to exclude DNET tumors.

```{r}
exclude_path_dx <- str_to_lower(
  c(
    "Dysembryoplastic neuroepithelial tumor"
  ))
```

Let's see how we do with the strategy laid out above.

```{r}
filtered_on_dx_df <- histologies_df %>%
  filter(str_detect(str_to_lower(pathology_diagnosis),
                    paste0(include_path_dx, collapse = "|")),
         # Designed to remove DNET samples
         str_detect(str_to_lower(pathology_diagnosis),
                    paste0(exclude_path_dx, collapse = "|"),
                    negate = TRUE)) %>%
  select(Kids_First_Biospecimen_ID, 
         sample_id, 
         Kids_First_Participant_ID,
         pathology_diagnosis,
         pathology_free_text_diagnosis,
         integrated_diagnosis, 
         short_histology)

filtered_on_dx_df
```

Let's tally the values in `pathology_diagnosis` in this data frame.

```{r}
filtered_on_dx_df %>%
  count(pathology_diagnosis) %>%
  arrange(desc(n))
```

This looks as we expect.

## Session Info

```{r}
sessionInfo()
```
